+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- hide tm-balance on PageProposal and PagePreferences if signed out @faboweb
+
 ## [1.0.0-beta.14] - 2019-03-06
 
 ### Fixed